requirements.txt
# Core
langchain==0.3.1
langchain-community==0.3.1

langchain-core>=0.3.0,<0.4.0
langgraph>=0.2.30,<0.3.0
sentence-transformers==3.0.1
tiktoken==0.8.0
datasets==3.0.0
pydantic==2.9.0
PyYAML==6.0.2

# Vector DB
chromadb==0.5.5
weaviate-client==4.8.1

# LLM + Inference
torch==2.4.1
transformers==4.44.2
accelerate==1.0.0
bitsandbytes==0.44.0

vllm==0.6.1; platform_system != "Windows"

# API & Deployment
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.32.0
requests==2.31.0
python-dotenv==1.0.0

# Eval & Tracking
ragas>=0.2.0,<0.3.0
mlflow==2.16.2

# Testing & Utils
pytest==8.3.3
tqdm==4.66.5
numpy==1.26.4
json-repair==0.7.0

# Jupyter/Dev utils 
ipykernel==7.1.0
ipython==9.8.0